Set it so that a pressure value will only be sent from AlphaLiveEngine if it is different from the last pressure value. This was a potential issue when setting a minPressureValue greater than 0 and scaling the values.

+++ Source/Functionality Classes/AlphaLiveEngine.h
@@ -121,7 +121,7 @@ private:
     EliteControls *eliteControls;
     
     int recievedPad;
-    float recievedValue;
+    int recievedValue;
     float recievedVelocity;
     
     //midi output device
